requirements.txt
Django>=3.0.7,<3.0.8
djangorestframework>=3.11.0,<3.12.0
psycopg2>=2.8.5,<=2.9.0
flake8>=3.8.2<=3.8.3